+++ vNES/src/vnes/Mapper004.java
@@ -66,7 +66,7 @@ public class Mapper004 extends MapperDefault {
         super.mapperInternalStateSave(buf);
 
         // Version:
-        buf.putByte((short) 1);
+        buf.putByte((int) 1);
 
         // State:
         buf.putInt(command);
@@ -80,7 +80,7 @@ public class Mapper004 extends MapperDefault {
 
     }
 
-    public void write(int address, short value) {
+    public void write(int address, int value) {
 
         if (address < 0x8000) {
