+++ vNES/src/vnes/Mapper022.java
@@ -25,7 +25,7 @@ public class Mapper022 extends MapperDefault {
 
     }
 
-    public void write(int address, short value) {
+    public void write(int address, int value) {
 
         if (address < 0x8000) {
             super.write(address, value);
