+++ vNES/src/vnes/Mapper023.java
@@ -29,7 +29,7 @@ public class Mapper023 extends MapperDefault {
         reset();
     }
 
-    public void write(int address, short value) {
+    public void write(int address, int value) {
 
         if (address < 0x8000) {
             super.write(address, value);
