+++ vNES/src/vnes/Mapper140.java
@@ -40,7 +40,7 @@ public class Mapper140 extends MapperDefault {
 
     }
 
-    public void write(int address, short value) {
+    public void write(int address, int value) {
 
         if (address < 0x8000) {
             // Handle normally:
